@@ -23,18 +23,18 @@
   "private": true,
   "dependencies": {
     "@angular-mdc/web": "^5.1.1",
-    "@angular/animations": "~13.3.11",
-    "@angular/cdk": "~13.3.9",
-    "@angular/common": "~13.3.11",
-    "@angular/core": "~13.3.11",
+    "@angular/animations": "^14.2.12",
+    "@angular/cdk": "^13.3.9",
+    "@angular/common": "^14.2.12",
+    "@angular/core": "^14.2.12",
     "@angular/flex-layout": "^13.0.0-beta.38",
-    "@angular/forms": "~13.3.11",
-    "@angular/material": "~13.3.9",
-    "@angular/platform-browser": "~13.3.11",
-    "@angular/platform-browser-dynamic": "~13.3.11",
+    "@angular/forms": "^14.2.12",
+    "@angular/material": "^13.3.9",
+    "@angular/platform-browser": "^14.2.12",
+    "@angular/platform-browser-dynamic": "^14.2.12",
     "@angular/pwa": "^13.3.9",
-    "@angular/router": "~13.3.11",
-    "@angular/service-worker": "~13.3.11",
+    "@angular/router": "^14.2.12",
+    "@angular/service-worker": "^14.2.12",
     "@auth0/auth0-spa-js": "^1.21.0",
     "@bugsnag/js": "^7.16.5",
     "@bugsnag/plugin-angular": "^7.16.5",
@@ -73,16 +73,16 @@
     "zone.js": "^0.11.5"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~13.3.9",
+    "@angular-devkit/build-angular": "^14.2.10",
     "@angular-eslint/builder": "^13.5.0",
     "@angular-eslint/eslint-plugin": "^13.5.0",
     "@angular-eslint/eslint-plugin-template": "^13.5.0",
     "@angular-eslint/schematics": "^13.5.0",
     "@angular-eslint/template-parser": "^13.5.0",
-    "@angular/cli": "~13.3.9",
-    "@angular/compiler": "~13.3.11",
-    "@angular/compiler-cli": "~13.3.11",
-    "@angular/language-service": "~13.3.11",
+    "@angular/cli": "^14.2.10",
+    "@angular/compiler": "^14.2.12",
+    "@angular/compiler-cli": "^14.2.12",
+    "@angular/language-service": "^14.2.12",
     "@triviality/logger": "^1.1.6",
     "@types/auth0-js": "^9.14.6",
     "@types/events": "^3.0.0",
